Fix #NaN

+++ back/src/main/java/com/github/an0rakdev/talkgcpmamie/datas/CartRepository.java
@@ -40,7 +40,7 @@ public class CartRepository {
 	@Transactional
 	public boolean updateQuantityOf(String code, int newQuantity) {
 		Query q = em.createQuery("UPDATE Cart c SET c.quantity = :q WHERE c.code = :c");
-		q.setParameter("q", newQuantity);
+		q.setParameter("q", (double) newQuantity);
 		q.setParameter("c", code);
 		return q.executeUpdate() <= 0;
 	}
